requirements.txt
asgiref==3.8.1
confusable-homoglyphs==3.3.1
dj-database-url==2.2.0
Django==4.2.13
django-debug-toolbar==4.4.2
django-registration==3.4
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing-extensions==4.12.1
